authorization implemented

+++ src/api/notes/handler.js
@@ -7,8 +7,9 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload)
     const { title = 'untitled', body, tags } = request.payload
+    const { id: credentialId } = request.auth.credentials
 
-    const noteId = await this._service.addNote({ title, body, tags })
+    const noteId = await this._service.addNote({ title, body, tags, owner: credentialId })
 
     return h.response({
       status: 'success',
@@ -19,8 +20,9 @@ class NotesHandler {
     }).code(201)
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes()
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials
+    const notes = await this._service.getNotes(credentialId)
     return {
       status: 'success',
       data: {
@@ -31,6 +33,8 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params
+    const { id: credentialId } = request.auth.credentials
+    await this._service.verifyNoteOwner(id, credentialId)
     const note = await this._service.getNoteById(id)
     return {
       status: 'success',
@@ -43,7 +47,9 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload)
     const { id } = request.params
+    const { id: credentialId } = request.auth.credentials
 
+    await this._service.verifyNoteOwner(id, credentialId)
     await this._service.editNoteById(id, request.payload)
 
     return {
@@ -54,6 +60,9 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params
+    const { id: credentialId } = request.auth.credentials
+
+    await this._service.verifyNoteOwner(id, credentialId)
     await this._service.deleteNoteById(id)
 
     return {
